refactor(settings): split settings-panel.tsx into per-concern submodules
Replace src/skills/default-layout/settings-panel.tsx (690 LOC) with a
settings-panel/ directory:

- panel.tsx          - main SettingsPanel: modal frame, callbacks,
                       form sections (Section/Field helpers inline)
- state.ts           - SettingsState type + readSettingsState reader
- constants.ts       - BUILTIN_THEMES, ANSI_PREVIEW_KEYS
- hooks.ts           - useConfigSnapshot, useEffectiveConfig,
                       useScrollToSection (kept together; tight chain)
- model-picker.tsx   - ModelPicker dropdown + custom-mode input
- extensions-list.tsx - ExtensionsList row renderer
- index.ts           - barrel re-exporting SettingsPanel

Public surface unchanged. The single consumer
(src/skills/default-layout/index.ts line 37) imports
"./settings-panel" which now resolves to the directory's index.ts.
