+++ src/skills/default-layout/settings-panel/panel.tsx
@@ -3,119 +3,33 @@
 // is a click away via the "Open config.toml" button.
 //
 // State contract (`/settings` slice on the main state object):
-//   { open: boolean, pending: { ui?, agent?, shell? } | null }
+//   { open: boolean, focusSection: string | null,
+//     pending: Partial<AethonConfig> | null }
 //
-// `pending` mirrors the user's unsaved edits — the form binds form
-// controls to it via $ref-style optimistic writes, so the user sees
-// changes apply live. Save serializes `pending` and invokes the Tauri
-// `write_config` command; Cancel discards `pending`. The panel reads
-// the current config state via `getConfig()` on mount so the form
-// reflects what's actually on disk, not stale in-memory tab state.
+// The panel reads the current config state via `getConfig()` on mount
+// so the form reflects what's actually on disk, not stale in-memory
+// tab state.
 
-import { useEffect, useMemo, useState } from "react";
 import { invoke } from "@tauri-apps/api/core";
 import { openUrl } from "@tauri-apps/plugin-opener";
-import type { BuiltinComponentProps } from "../../components/A2UIRenderer";
-import { getConfig, type AethonConfig } from "../../config";
-import { SHARE_MODES, type ShareMode } from "../../utils/shareMode";
-import { DropdownPickerCore } from "./variation-components";
-import { ToggleSwitch } from "./toggle-switch";
-
-interface SettingsState {
-  open: boolean;
-  focusSection: string | null;
-  /** User's unsaved edits. Null when the panel hasn't loaded the
-   *  config yet OR the user hasn't touched anything. The form reads
-   *  from `pending` first, falling back to the config snapshot. */
-  pending: Partial<AethonConfig> | null;
-}
-
-function readSettingsState(state: Record<string, unknown>): SettingsState {
-  const s = (state.settings as Partial<SettingsState> | undefined) ?? {};
-  return {
-    open: !!s.open,
-    focusSection:
-      typeof s.focusSection === "string" ? s.focusSection : null,
-    pending: (s.pending as Partial<AethonConfig> | null) ?? null,
-  };
-}
-
-const BUILTIN_THEMES = [
-  { id: "ember", label: "Ember — warm dark" },
-  { id: "paper", label: "Paper — cream light" },
-  { id: "aether", label: "Æther — signature" },
-  { id: "brink", label: "Brink — Ristretto warm chrome with gold accent" },
-];
-
-// 16-swatch ANSI preview block — order matches xterm's standard ANSI
-// indices (0..7 standard + 8..15 bright) so users reading a terminal
-// `\x1b[34m` recognise the swatch position. Bound to CSS vars so the
-// preview tracks the active theme live.
-const ANSI_PREVIEW_KEYS = [
-  "--ansi-black",
-  "--ansi-red",
-  "--ansi-green",
-  "--ansi-yellow",
-  "--ansi-blue",
-  "--ansi-magenta",
-  "--ansi-cyan",
-  "--ansi-white",
-  "--ansi-bright-black",
-  "--ansi-bright-red",
-  "--ansi-bright-green",
-  "--ansi-bright-yellow",
-  "--ansi-bright-blue",
-  "--ansi-bright-magenta",
-  "--ansi-bright-cyan",
-  "--ansi-bright-white",
-] as const;
+import type { BuiltinComponentProps } from "../../../components/A2UIRenderer";
+import type { AethonConfig } from "../../../config";
+import { SHARE_MODES, type ShareMode } from "../../../utils/shareMode";
+import { ANSI_PREVIEW_KEYS, BUILTIN_THEMES } from "./constants";
+import { ExtensionsList } from "./extensions-list";
+import {
+  useConfigSnapshot,
+  useEffectiveConfig,
+  useScrollToSection,
+} from "./hooks";
+import { ModelPicker } from "./model-picker";
+import { readSettingsState } from "./state";
 
 export function SettingsPanel({ state, onEvent }: BuiltinComponentProps) {
   const settings = readSettingsState(state);
-
-  // Live config snapshot. Loaded via getConfig() on first open of
-  // the panel and refreshed on save round-trips. Held in state (not
-  // a ref) so the form re-renders when the value lands.
-  const [configSnapshot, setConfigSnapshot] = useState<AethonConfig | null>(
-    null,
-  );
-  useEffect(() => {
-    if (!settings.open) return;
-    let cancelled = false;
-    void getConfig().then((cfg) => {
-      if (!cancelled) setConfigSnapshot(cfg);
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [settings.open]);
-
-  // Effective config = pending overlay over live config snapshot.
-  // Form bindings read from this; Save / cancel manipulate `pending`.
-  const eff = useMemo<AethonConfig | null>(() => {
-    if (!configSnapshot) return null;
-    const p = settings.pending ?? {};
-    return {
-      ui: { ...configSnapshot.ui, ...(p.ui ?? {}) },
-      agent: { ...configSnapshot.agent, ...(p.agent ?? {}) },
-      shell: { ...configSnapshot.shell, ...(p.shell ?? {}) },
-      shortcuts: { ...configSnapshot.shortcuts, ...(p.shortcuts ?? {}) },
-    };
-  }, [configSnapshot, settings.pending]);
-
-  useEffect(() => {
-    if (!settings.open || !eff || !settings.focusSection) return;
-    const frame = window.requestAnimationFrame(() => {
-      const target = [
-        ...document.querySelectorAll<HTMLElement>("[data-settings-section]"),
-      ].find((el) => el.dataset.settingsSection === settings.focusSection);
-      target?.scrollIntoView({
-        block: "start",
-        behavior: "smooth",
-      });
-    });
-    return () => window.cancelAnimationFrame(frame);
-  }, [eff, settings.focusSection, settings.open]);
+  const configSnapshot = useConfigSnapshot(settings.open);
+  const eff = useEffectiveConfig(configSnapshot, settings.pending);
+  useScrollToSection(settings.open, eff, settings.focusSection);
 
   if (!settings.open) return null;
 
@@ -470,221 +384,5 @@ function Field(props: { label: string; children: React.ReactNode }) {
       <span className="ae-settings-field-label">{props.label}</span>
       <span className="ae-settings-field-control">{props.children}</span>
     </label>
-  );
-}
-
-interface ModelOption {
-  id: string;
-  label: string;
-}
-
-/**
- * Select-with-custom-fallback model picker. Backed by the live model
- * list at `state.sidebar.models` (populated by the bridge's `ready` +
- * model_changed events). Two interaction modes:
- *
- *  - Dropdown of the loaded models, plus "Other…" which swaps in a free
- *    text input so power users can still type "anthropic/claude-foo".
- *  - Custom mode is selected automatically when the saved value
- *    doesn't match any loaded model (e.g. user changed their pi config
- *    but hasn't relaunched the bridge yet).
- *
- * Empty value (no entry / null in config) renders the placeholder
- * "(pi default)" — the agent picks based on env vars in that case.
- */
-function ModelPicker({
-  state,
-  value,
-  onChange,
-}: {
-  state: Record<string, unknown>;
-  value: string;
-  onChange: (next: string) => void;
-}) {
-  // Reuses the same DropdownPickerCore the header model picker renders,
-  // so search/filter + keyboard nav + visual style stay in lockstep
-  // (replaces a bespoke <select>+custom-mode pair). A sentinel
-  // `__pi_default__` id maps back to `""` so "(pi default)" stays
-  // selectable; a sentinel `__custom__` id flips into a free-text
-  // input for the rare case a user wants a model not on the list.
-  const sidebar =
-    (state.sidebar as Record<string, unknown> | undefined) ?? {};
-  const models =
-    ((sidebar.models as ModelOption[] | undefined) ?? []).filter(
-      (m) => typeof m.id === "string" && m.id.length > 0,
-    );
-  const trimmed = value.trim();
-  const knownIds = new Set(models.map((m) => m.id));
-  const startsCustom = trimmed.length > 0 && !knownIds.has(trimmed);
-  const [customMode, setCustomMode] = useState(startsCustom);
-  // Re-flip when the value or model list changes (e.g. an extension
-  // registers a new model). Derived-state resync from external input.
-  useEffect(() => {
-    // eslint-disable-next-line react-hooks/set-state-in-effect -- derived-state resync on external input change
-    setCustomMode(startsCustom);
-  }, [startsCustom]);
-
-  if (customMode) {
-    return (
-      <span className="ae-settings-model-picker">
-        <input
-          type="text"
-          className="ae-settings-input"
-          placeholder="anthropic/claude-sonnet-4-6"
-          value={trimmed}
-          onChange={(e) => onChange(e.target.value)}
-          aria-label="Custom model id"
-        />
-        <button
-          type="button"
-          className="ae-settings-secondary ae-settings-model-picker-toggle"
-          onClick={() => {
-            setCustomMode(false);
-            onChange("");
-          }}
-          title="Pick from the loaded model list"
-        >
-          List
-        </button>
-      </span>
-    );
-  }
-
-  const PI_DEFAULT = "__pi_default__";
-  const CUSTOM = "__custom__";
-  const activeMatch = knownIds.has(trimmed) ? trimmed : PI_DEFAULT;
-  const items = [
-    { id: PI_DEFAULT, label: "(pi default — picks from env vars)", active: activeMatch === PI_DEFAULT },
-    ...models.map((m) => ({
-      id: m.id,
-      label: m.label || m.id,
-      hint: m.label && m.label !== m.id ? m.id : undefined,
-      active: activeMatch === m.id,
-    })),
-    { id: CUSTOM, label: "Custom id…" },
-  ];
-  const activeItem = items.find((it) => it.active);
-  const buttonLabel = activeItem?.label || "(pi default)";
-
-  return (
-    <span className="ae-settings-model-picker">
-      <DropdownPickerCore
-        className="a2ui-model-picker ae-settings-model-picker-dropdown"
-        buttonLabel={buttonLabel}
-        align="left"
-        sections={[
-          {
-            id: "models",
-            title: "models",
-            items,
-            searchable: true,
-            searchPlaceholder: "filter models — sonnet, gpt, qwen…",
-            emptyLabel: "no models match",
-          },
-        ]}
-        onSelect={(_sectionId, itemId) => {
-          if (itemId === CUSTOM) {
-            setCustomMode(true);
-            return;
-          }
-          onChange(itemId === PI_DEFAULT ? "" : itemId);
-        }}
-      />
-    </span>
-  );
-}
-
-interface ExtensionItem {
-  id: string;
-  label: string;
-  hint?: string;
-  active?: boolean;
-}
-
-function ExtensionsList({
-  state,
-  onEvent,
-}: {
-  state: Record<string, unknown>;
-  onEvent: BuiltinComponentProps["onEvent"];
-}) {
-  const sidebar = (state.sidebar as Record<string, unknown> | undefined) ?? {};
-  const items = (sidebar.extensions as ExtensionItem[] | undefined) ?? [];
-  if (items.length === 0) {
-    return (
-      <p className="ae-settings-note">
-        No user or project extensions loaded. Drop a <code>.mjs</code> into{" "}
-        <code>~/.aethon/extensions/</code> or the active project's{" "}
-        <code>.aethon/extensions/</code> directory to register one.
-      </p>
-    );
-  }
-  return (
-    <ul className="ae-settings-ext-list">
-      {items.map((item) => {
-        const kind = item.id.startsWith("ext:")
-          ? "enabled"
-          : item.id.startsWith("ext-failed:")
-            ? "failed"
-            : item.id.startsWith("ext-disabled:")
-              ? "disabled"
-              : "core";
-        const name =
-          kind === "enabled"
-            ? item.id.slice("ext:".length)
-            : kind === "failed"
-              ? item.id.slice("ext-failed:".length)
-              : kind === "disabled"
-                ? item.id.slice("ext-disabled:".length)
-                : item.label;
-        const canToggle = kind !== "core";
-        return (
-          <li
-            key={item.id}
-            className={`ae-settings-ext-row ae-settings-ext-row--${kind}`}
-          >
-            <span className="ae-settings-ext-name">{item.label}</span>
-            {item.hint ? (
-              <span className="ae-settings-ext-hint">{item.hint}</span>
-            ) : null}
-            {canToggle ? (
-              <ToggleSwitch
-                checked={kind === "enabled"}
-                disabled={kind === "failed"}
-                // Pick the verb from the actual checked state (not just
-                // the explicit "disabled" branch) — otherwise a failed
-                // extension renders unchecked but announces "Disable
-                // extension X" to assistive tech. Matches the sidebar
-                // toggle which keys off `extState.checked` for the
-                // same reason.
-                ariaLabel={`${kind === "enabled" ? "Disable" : "Enable"} extension ${name}`}
-                title={
-                  kind === "failed"
-                    ? "Extension failed to load — fix the error and reload to re-enable"
-                    : kind === "disabled"
-                      ? `Enable ${name}`
-                      : `Disable ${name}`
-                }
-                onChange={(next) => {
-                  onEvent("toggle-extension", {
-                    sectionId: "extensions",
-                    itemId: item.id,
-                    name,
-                    disabled: !next,
-                  });
-                }}
-              />
-            ) : (
-              <span
-                className="ae-settings-ext-hint"
-                title="Built-in core extension"
-              >
-                core
-              </span>
-            )}
-          </li>
-        );
-      })}
-    </ul>
   );
 }
